+++ NEWS.md
@@ -8,6 +8,12 @@ Cite all releases with doi [10.5281/zenodo.5226085](https://doi.org/10.5281/zeno
 which always resolves to the latest release.
 
 
+* New function `read_ecc_from_excel()` serves
+  as the inverse of `write_ecc_to_excel()`, 
+  taking advantage of named regions in Excel spreadsheets.
+  Both functions now name regions as
+  `<<matrix name>>__<<worksheet name>>` by default.
+
 
 ## Recca 0.1.59 (2025-06-27) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15756610.svg)](https://doi.org/10.5281/zenodo.15756610)